1. Поправил обработчик события клика по игровой карте; 2. Поправил несколько косяков в классе логики бойцов; 3. Поправил баг сериализации в классе ServerMoveWarriorAnswer; 4. Косметические изменения в коде других классов.

+++ server.battle/src/com/geargames/regolith/serializers/answers/ServerMoveWarriorAnswer.java
@@ -69,7 +69,7 @@ public class ServerMoveWarriorAnswer extends SerializedMessage {
                     SimpleSerializer.serialize(warrior.getCellY(), buffer);
                 }
             } else {
-                SimpleSerializer.serialize(0, buffer);
+                SimpleSerializer.serialize((byte) 0, buffer);
             }
         }
     }
